[DQMOffline/Trigger] removed plugins/TopMonitor.h
